Criação de exception

+++ app/Controllers/Http/CategoryController.ts
@@ -1,20 +1,24 @@
 import type { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
+
 import Category from 'App/Models/Category'
+
 import CategoryValidator from 'App/Validators/CategoryValidator'
+import CategoryException from 'App/Exceptions/CategoryException'
+
 
 export default class CategoryController {
-  
+
   public async index() {
 
-    const categories = await Category.all()
+    const categories = await Category.query().orderBy('id', 'desc');
 
     return categories
   }
 
   public async store({ request, auth, response }: HttpContextContract) {
-  
+
     const categoryData = await request.validate(CategoryValidator)
-    
+
     const user = await auth.authenticate()
 
     const category = await user.related('category').create(categoryData)
@@ -25,14 +29,14 @@ export default class CategoryController {
   }
 
   public async show({ params }: HttpContextContract) {
-    
+
     const category = await Category.findOrFail(params.id)
 
     return category
 
   }
 
-  public async update({ params, request, response  }: HttpContextContract) {
+  public async update({ params, request, response }: HttpContextContract) {
 
     const categoryData = await request.validate(CategoryValidator)
 
@@ -49,11 +53,16 @@ export default class CategoryController {
     return category
   }
 
-  public async destroy({ params,  response }: HttpContextContract) {
+
+  public async destroy({ params, response }: HttpContextContract) {
 
     const category = await Category.findOrFail(params.id)
 
-    await category.related('question').detach()
+    const questionsCount = await category.related('question').query()
+
+    if (questionsCount.length > 0) {
+      throw new CategoryException('This category has questions attached and cannot be deleted.', 400);
+    }
 
     await category.delete()
